Add MIC field calculation
NTLMv2 `authenticate` messages are expected to include MIC field.
MIC is a hash of 3 concatenated exchanged messages (`negotiate`, `challenge` and `authenticate`)
As a MIC field is part of the `authenticate` message, during a calculation the field is expected to be filled out with zeroes.
Once calculated the value is back-filled inside the `authenticate` byte array.